+++ src/components/MentionTextField/mention-types.d.ts
@@ -1,7 +1,8 @@
-import React, { CSSProperties, FunctionComponent } from "react";
+import { CSSProperties, FunctionComponent, ReactNode } from "react";
 import { BaseEditor, BaseRange, Descendant, Editor } from "slate";
 import { HistoryEditor } from "slate-history";
 import { ReactEditor } from "slate-react";
+import { WithChildren } from "../../types/common";
 
 interface MentionTextFieldState {
   editor: Editor;
@@ -16,7 +17,7 @@ interface MentionTextFieldProps
   placeholder?: string;
   initialValue?: string;
   onChange?: (value: string) => void;
-  addMentionText?: (value: string) => string;
+  addMentionText?: (value: string) => ReactNode;
   onEnterPress?: () => void;
   suggestionPopoverZIndex?: number;
   suggestionListComponent?: FunctionComponent;
@@ -28,7 +29,7 @@ interface MentionTextFieldHookOptions {
   singleLine?: boolean;
 }
 
-interface SuggestionListItemProps {
+interface SuggestionListItemProps extends WithChildren {
   onClick: () => void;
   selected: boolean;
 }