+++ src/components/Onboarding.tsx
@@ -3,6 +3,7 @@ import FolderOpenOutlinedIcon from "@mui/icons-material/FolderOpenOutlined";
 import { Box, Button, Stack, styled, Typography } from "@mui/material";
 import { useTranslation } from "react-i18next";
 import { useFileCreateDialog } from "../data/FileCreateDialogContext";
+import { useLoading } from "../data/LoadingContext";
 import { useTask } from "../data/TaskContext";
 import logo from "../images/logo.png";
 import { usePlatform } from "../utils/platform";
@@ -24,11 +25,14 @@ const StyledBox = styled("div")`
 const Onboarding = () => {
   const { t } = useTranslation();
   const platform = usePlatform();
+  const { loading } = useLoading();
   const { setFileCreateDialog } = useFileCreateDialog();
   const { taskLists, openTodoFilePicker } = useTask();
 
   return (
-    <StyledBox sx={{ display: taskLists.length === 0 ? "flex" : "none" }}>
+    <StyledBox
+      sx={{ display: !loading && taskLists.length === 0 ? "flex" : "none" }}
+    >
       <Stack spacing={1}>
         <Box sx={{ py: 1, textAlign: "center" }}>
           <img src={logo} alt="Logo" height={96} style={{ opacity: 0.2 }} />