+++ src/types/common.ts
@@ -1,9 +1,5 @@
 import { ReactNode } from "react";
 
-export interface Dictionary<T> {
-  [key: string]: T;
-}
-
 export interface WithChildren {
   children?: ReactNode;
 }