+++ src/utils/testing.tsx
@@ -3,7 +3,6 @@ import { Filesystem, ReadFileResult } from "@capacitor/filesystem";
 import { ReadFileOptions } from "@capacitor/filesystem/dist/esm/definitions";
 import { GetOptions, GetResult, Storage } from "@capacitor/storage";
 import i18n from "i18next";
-import { PropsWithChildren } from "react";
 import { initReactI18next } from "react-i18next";
 import { MemoryRouter } from "react-router-dom";
 import { vi } from "vitest";
@@ -48,7 +47,7 @@ export interface SecureStorageItem {
   value: string;
 }
 
-export interface TestContextProps {
+export interface TestContextProps extends WithChildren {
   text?: string;
   filesystem?: FilesystemItem[];
   storage?: StorageItem[];
@@ -178,9 +177,7 @@ const Page = ({ children }: WithChildren) => {
   return <div data-testid={loading ? "loading" : "page"}>{children}</div>;
 };
 
-export const EmptyTestContext = (
-  props: PropsWithChildren<TestContextProps>
-) => {
+export const EmptyTestContext = (props: TestContextProps) => {
   const { text, filesystem, storage, secureStorage, platform, children } =
     props;